requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dask == 2022.6.1
gym == 0.19.0
mnist == 0.2.2
numpy == 1.19.5
pandas == 1.2.2
pygame == 2.0.1
scipy == 1.4.1
stable_baselines3 == 1.2.1a1
tensorflow_gpu == 2.4.1
torch == 1.9.0
treelib == 1.6.1
